+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Added
+- Ability to have custom skin for items embed, use `data-armory-{itemId}-skin="{skinId}"`
 
 ## [4.37.2][] - 2017-09-13
 ### Fixed